🧹 fix circular deps
