+++ packages/curvenote-cli/src/session/index.ts
@@ -1,4 +1,4 @@
 export * from './session.js';
 export * from './auth.js';
 export * from './types.js';
-export * from './utils.js';
+export * from './utils/index.js';
