+++ packages/curvenote-cli/src/session/utils/getConfigPath.ts
@@ -1,0 +1,17 @@
+import path from 'node:path';
+
+export function getConfigPath() {
+  const pathArr: string[] = [];
+  const local = ['curvenote', 'settings.json'];
+  if (process.env.APPDATA) {
+    pathArr.push(process.env.APPDATA);
+  } else if (process.platform === 'darwin' && process.env.HOME) {
+    pathArr.push(path.join(process.env.HOME, '.config'));
+  } else if (process.env.HOME) {
+    pathArr.push(process.env.HOME);
+    if (local.length > 0) {
+      local[0] = `.${local[0]}`;
+    }
+  }
+  return path.join(...pathArr, ...local);
+}
