+++ packages/curvenote-cli/src/session/utils/getLogLevel.ts
@@ -1,0 +1,10 @@
+import { LogLevel } from 'myst-cli-utils';
+
+/**
+ * Duplicated from myst-cli-utils, where function is not exported
+ */
+export function getLogLevel(level: LogLevel | boolean | string = LogLevel.info): LogLevel {
+  if (typeof level === 'number') return level;
+  const useLevel: LogLevel = level ? LogLevel.debug : LogLevel.info;
+  return useLevel;
+}
